\name{dependencies-package}
\alias{dependencies-package}
\alias{dependencies}
\docType{package}
\title{
\packageTitle{dependencies}
}
\description{
\packageDescription{dependencies}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{dependencies}
\packageIndices{dependencies}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{dependencies}

Maintainer: \packageMaintainer{dependencies}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
# simple examples of the most important functions
}
